+++ src/TrustedSigning/TrustedSigning/ChangeLog.md
@@ -18,9 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated InvokeCIPolicySigning to support signing files with the .cip extension
 
 ## Version 0.1.1
-* Modified InvokeCiPolicySigning to include ShouldProcess command confirmation
+* Modified InvokeCIPolicySigning to include ShouldProcess command confirmation
 
 ## Version 0.1.0
 * Renamed from Az.CodeSigning